+++ src/analyzers/SourceImpl.ts
@@ -13,21 +13,33 @@ class SourceImpl implements Source {
   private readonly lines: string[]
 
   constructor(source: string) {
-    this.lines = source.split("\n")
+    this.lines = source.split(/\r?\n/)
   }
 
-  public get(node: NodeWithLocation): string {
-    const start = this.lines[node.loc.start.line - 1]
-    const end = this.lines[node.loc.end.line - 1]
-    if (start === end) {
-      return start.substring(node.loc.start.column, node.loc.end.column)
+  public getLines(node: NodeWithLocation): string[] {
+    const startLineLoc = Math.max(0, node.loc.start.line - 1)
+    const start = this.lines[startLineLoc]
+    const endLineLoc = Math.min(this.lines.length - 1, node.loc.end.line - 1)
+
+    if (startLineLoc === endLineLoc) {
+      return [start.substring(node.loc.start.column, node.loc.end.column)]
     }
+
+    const end = this.lines[endLineLoc]
 
     return [
       start.substring(node.loc.start.column),
-      ...this.lines.slice(node.loc.start.line, node.loc.end.line - 2),
-      end.substring(0, node.loc.end.column)
-    ].join("\n")
+      ...(
+        startLineLoc + 1 <= endLineLoc - 1
+          ? this.lines.slice(startLineLoc + 1, endLineLoc - 1)
+          : []
+      ),
+      end.substring(0, node.loc.end.column < 0 ? undefined : node.loc.end.column)
+    ]
+  }
+
+  public get(node: NodeWithLocation): string {
+    return this.getLines(node).join("\n")
   }
 
   public getOuter(node: NodeWithLocation): string {